frameworks/observability/requirements.txt
pyyaml==6.0.2
jsonschema==4.22.0
click==8.1.7
ruff==0.6.9
mypy==1.11.2
rich==13.7.1
tabulate==0.9.0
types-PyYAML==6.0.12.20240808
